% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degCreCore.R
\docType{package}
\name{DegCre}
\alias{DegCre}
\alias{DegCre-package}
\title{DegCre}
\description{
Probabilistic association of DEGs to CREs from differential data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brianSroberts/DegCre}
  \item Report bugs at \url{https://github.com/brianSroberts/DegCre/issues}
}

}
\author{
\strong{Maintainer}: Brian S. Roberts \email{brianroberts1976@yahoo.com} (\href{https://orcid.org/0009-0001-2914-6826}{ORCID})

}
