% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degCreCore.R
\docType{data}
\name{DexNR3C1}
\alias{DexNR3C1}
\title{DegCre input data for examples.}
\description{
DegCre input data for examples.
}
\details{
This is a list with two slots: \code{DegGR} and \code{CreGR}.
This data was derived from work by McDowell et al. in which they generated RNA-seq and ChIP-seq data by treating A549 cells with dexamethasone at several time points. Specifically this is RNA-seq and NR3C1 ChIP-seq at four hours versus control.
}
\references{
\url{https://genome.cshlp.org/content/28/9/1272}
}
\author{
Brian S. Roberts
}
\keyword{data}
