% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degCreCore.R
\docType{package}
\name{DegCre}
\alias{DegCre}
\alias{_PACKAGE}
\alias{DegCre-package}
\title{DegCre}
\description{
Probabilistic association of DEGs to CREs from differential data.
}
\author{
\strong{Maintainer}: Brian S. Roberts \email{broberts@hudsonalpha.org}

}
